Order Managerment

+++ db/migrate/20180928162504_create_orders.rb
@@ -2,7 +2,7 @@ class CreateOrders < ActiveRecord::Migration[5.2]
   def change
     create_table :orders do |t|
       t.integer :status, null: false, default: 0
-      t.integer :total_cost
+      t.decimal :total_cost
       t.references :user, foreign_key: true
 
       t.timestamps
