+++ db/migrate/20180928163427_create_products.rb
@@ -3,7 +3,7 @@ class CreateProducts < ActiveRecord::Migration[5.2]
     create_table :products do |t|
       t.string :name
       t.text :description
-      t.integer :price, default: 0
+      t.decimal :price, default: 0
       t.integer :average_rate, default: 0
       t.references :category, foreign_key: true
